+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Changelog
 - Remember where danbooru files where downloaded too
 - Be able to group the search result again with group=SIZE
 - Add danbooru API limit warning
+- Allow tag options and qualifiers like ``-some_tag``, ``*some`` & ``order:score``
 
 
 2.3.0 (2019-01-05)